* Refactor * Add docker file

+++ src/main/java/com/example/roomoccupancymanager/config/RoomPriceConfig.java
@@ -3,9 +3,11 @@ package com.example.roomoccupancymanager.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.ConstructorBinding;
 
+import java.math.BigDecimal;
+
 @ConfigurationProperties("hotel.room-prices")
 @ConstructorBinding
 public record RoomPriceConfig(
-        Double premiumPriceThreshold
+        BigDecimal premiumPriceThreshold
 ) {
 }
